fix(search): use the bundled ripgrep in ElectronFileSystemService
`searchFiles` spawned a bare `rg`, so it depended on the user having ripgrep
installed globally and failed with ENOENT when they did not — even though
Nimbalyst ships its own binary and WorkspaceHandlers already resolved it
correctly through a private helper.

Extract that helper into `services/ripgrepPath.ts` and use it from both call
sites, so there is one way to find the binary rather than one correct way and
one bare string.
